worker: add archive.go
