+++ worker/archive.go
@@ -1,0 +1,35 @@
+package worker
+
+import (
+	"archive/tar"
+	"bytes"
+	"io/ioutil"
+	"path"
+)
+
+func archive(src string) (*bytes.Reader, error) {
+	buf := new(bytes.Buffer)
+	tw := tar.NewWriter(buf)
+
+	content, err := ioutil.ReadFile(src)
+	if err != nil {
+		return nil, err
+	}
+
+	h := &tar.Header{
+		Name: path.Base(src),
+		Mode: 0755,
+		Size: int64(len(content)),
+	}
+	if err := tw.WriteHeader(h); err != nil {
+		return nil, err
+	}
+	if _, err := tw.Write(content); err != nil {
+		return nil, err
+	}
+	if err := tw.Close(); err != nil {
+		return nil, err
+	}
+
+	return bytes.NewReader(buf.Bytes()), nil
+}
