+++ worker/worker.go
@@ -2,16 +2,13 @@ package worker
 
 import (
 	"archive/tar"
-	"bytes"
 	log "github.com/Sirupsen/logrus"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/client"
 	"golang.org/x/net/context"
 	"io"
-	"io/ioutil"
 	"os"
-	"path"
 	"strings"
 )
 
@@ -48,33 +45,6 @@ func (w *Worker) Create(ctx context.Context, imageName string, cmd string) (err 
 
 	w.id = c.ID
 	return nil
-}
-
-func archive(src string) (*bytes.Reader, error) {
-	buf := new(bytes.Buffer)
-	tw := tar.NewWriter(buf)
-
-	content, err := ioutil.ReadFile(src)
-	if err != nil {
-		return nil, err
-	}
-
-	h := &tar.Header{
-		Name: path.Base(src),
-		Mode: 0755,
-		Size: int64(len(content)),
-	}
-	if err := tw.WriteHeader(h); err != nil {
-		return nil, err
-	}
-	if _, err := tw.Write(content); err != nil {
-		return nil, err
-	}
-	if err := tw.Close(); err != nil {
-		return nil, err
-	}
-
-	return bytes.NewReader(buf.Bytes()), nil
 }
 
 func (w *Worker) Copy(ctx context.Context, tar io.Reader, dst string) error {
